+++ src/graphql/account/account.query.input.type.ts
@@ -1,12 +1,12 @@
-import { Field, InputType, Int } from "@nestjs/graphql";
+import { Field, InputType, Float } from "@nestjs/graphql";
 import { QueryPagination } from "src/common/entities/query.pagination";
 
 @InputType({ description: "Get accounts input." })
 export class GetAccountsInput {
-  @Field(() => Int, { name: "from", description: "Number of accounts to skip for the given result set.", nullable: true, defaultValue: 0 })
+  @Field(() => Float, { name: "from", description: "Number of accounts to skip for the given result set.", nullable: true, defaultValue: 0 })
   skip: number = 0;
 
-  @Field(() => Int, { name: "size", description: "Number of accounts to retrieve for the given result set.", nullable: true, defaultValue: 25 })
+  @Field(() => Float, { name: "size", description: "Number of accounts to retrieve for the given result set.", nullable: true, defaultValue: 25 })
   size: number = 25;
 
   public static resolve(input: GetAccountsInput): QueryPagination {
